+++ backend/routes copy.py
@@ -1,0 +1,38 @@
+from flask import Flask, request, jsonify, session, redirect
+from backend.models import db, User
+from flask_cors import CORS
+
+app = Flask(__name__)
+app.secret_key = 'your_secret_key'
+CORS(app)
+
+@app.route('/signup', methods=['POST'])
+def signup():
+    data = request.json
+    new_user = User(email=data['email'], password=data['password'])
+    db.session.add(new_user)
+    db.session.commit()
+    return jsonify({"message": "User created successfully"}), 201
+
+@app.route('/login', methods=['POST'])
+def login():
+    data = request.json
+    user = User.query.filter_by(email=data['email']).first()
+    if user and user.password == data['password']:
+        session['user_id'] = user.id
+        return jsonify({"message": "Login successful"})
+    return jsonify({"message": "Invalid credentials"}), 401
+
+@app.route('/private')
+def private():
+    if 'user_id' in session:
+        return jsonify({"message": "Welcome to the private dashboard"})
+    return redirect('/login')
+
+@app.route('/logout')
+def logout():
+    session.pop('user_id', None)
+    return jsonify({"message": "Logged out successfully"}), 200
+
+if __name__ == '__main__':
+    app.run(debug=True)
